pyproject.toml
[project]
name = "rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "beam-client>=0.2.97",
    "ruff>=0.7.1",
    "vllm>=0.6.3.post1",
]
